services/user_node/requirements.txt
pycryptodome
requests
psycopg2-binary # just for development, exclude later 
flask # just for development, exclude later